% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_fct.R
\name{add_overlayers}
\alias{add_overlayers}
\title{Add all the overlayers}
\usage{
add_overlayers(map, overlayers)
}
\arguments{
\item{map}{leaflet map}

\item{overlayers}{overlayers data.frame}
}
\value{
overlayers for leaflet
}
\description{
Add all the overlayers
}
\examples{
map \%>\%
add_overlayers(data_overlayers_df())
}
