% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data.R
\name{assign_classes}
\alias{assign_classes}
\title{Assign classes to network dgos}
\usage{
assign_classes(data, classes)
}
\arguments{
\item{data}{dataframe or sf object to which classes}

\item{variables}{vector of variable names on which the classification is based}

\item{greater_thans}{vector of class-thresholds for classification}

\item{class_names}{vector containing the names of all classes to be assigned}
}
\value{
classified dataframe/sf object with additional variable: class
}
\description{
Assign classes to network dgos
}
\examples{
classified_network <- network_dgo \%>\%
    assign_classes(variables = as.character(r_val$grouping_table_data$variable),
    greater_thans = r_val$grouping_table_data$greaterthan,
    class_names = r_val$grouping_table_data$class)

}
