% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data.R
\name{assign_classes_manual}
\alias{assign_classes_manual}
\title{Assign classes to network dgos}
\usage{
assign_classes_manual(data, classes)
}
\arguments{
\item{data}{dataframe or sf object which contains dgos of axis or region}

\item{classes}{df containing columns variables, greater_thans, class_names, colors which define the classification of the network}
}
\value{
classified dataframe/sf object with additional variables: class_name and color
}
\description{
Assign classes to network dgos
}
\examples{
classified_network <- network_dgo \%>\%
    assign_classes(variables = as.character(r_val$grouping_table_data$variable),
    greater_thans = r_val$grouping_table_data$greaterthan,
    class_names = r_val$grouping_table_data$class)

}
