% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_bassin_hydrographique.R
\docType{data}
\name{bassin_hydrographique}
\alias{bassin_hydrographique}
\title{bassin_hydrographique}
\format{
A data frame with 1 rows and 6 variables:
\describe{
  \item{ gid }{  numeric }
  \item{ cdbh }{  character }
  \item{ lbbh }{  character }
  \item{ numcircadm }{  character }
  \item{ projcoordo }{  character }
  \item{ geom }{  sfc_MULTIPOLYGON,sfc }
}
}
\source{
Source
}
\usage{
bassin_hydrographique
}
\description{
Description.
}
\keyword{datasets}
