% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_analysis.R
\name{classification_scales_overview_table}
\alias{classification_scales_overview_table}
\title{Create an overview table of the value-distribution of a variable for each scale (region and axis)}
\usage{
classification_scales_overview_table(data, var)
}
\arguments{
\item{data}{classified network data with entries for regional and axis dgos}

\item{var}{variable for which table should be created}
}
\value{
DT-object with stats calculated for one variable for each scale-group
}
\description{
Create an overview table of the value-distribution of a variable for each scale (region and axis)
}
\examples{
classification_scales_overview_table(data_plots, "forest_pc")
}
