% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_cr_profile.R
\name{cr_profile_empty}
\alias{cr_profile_empty}
\title{Create an empty cross section plot.}
\usage{
cr_profile_empty()
}
\value{
An empty longitudinal profile plot with a specified title.
}
\description{
This function generates an empty cross section plot using the 'plot_ly'
function from the 'plotly' package.
}
\examples{
# Create an empty longitudinal profile plot
empty_plot <- lg_profile_empty()
empty_plot

}
