% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_cr_profile.R
\name{cr_section_main}
\alias{cr_section_main}
\title{Create a cross section plot for selected DGO data.}
\usage{
cr_section_main(data)
}
\arguments{
\item{data}{data.frame elevation profiles from selected dgo.}
}
\value{
plotly cross section plot.
}
\description{
Create a cross section plot for selected DGO data.
}
