% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_analysis_tables.R
\name{create_analysis_table}
\alias{create_analysis_table}
\title{Create reactable table for metrics-statistics}
\usage{
create_analysis_table(df, vars, scale_name = "")
}
\arguments{
\item{df}{metric-statistics dataframe with *_distr-columns}

\item{vars}{metric names to be included in table}

\item{strahler_sel}{selected Strahler order, default is 0 which represents an aggregate of the whole entity}
}
\value{
reactable table with variables and sparklines
}
\description{
Create reactable table for metrics-statistics
}
\examples{
create_analysis_table(df, vars = c("crops_pc", "dense_urban_pc", "dense_urban"))
}
