% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot_long.R
\name{create_classes_background}
\alias{create_classes_background}
\title{Create background-layout of  classes for longitudinal profile}
\usage{
create_classes_background(classified_axis)
}
\arguments{
\item{classified_axis}{}
}
\value{
list ready to be used for plotly::layout()-function
}
\description{
Create background-layout of  classes for longitudinal profile
}
