% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_classes_manual.R
\name{create_df_input}
\alias{create_df_input}
\title{Create initial dataframe for 1-variable classification, to be displayed on the UI table}
\usage{
create_df_input(variable_name, q_0025, q_0975, quantile = 95, no_classes = 4)
}
\arguments{
\item{variable_name}{name of variable for which the classification should be undertaken}

\item{q_0025}{2.5\%-quantile value of selected metric}

\item{q_0975}{97.5\%-quantile value of selected metric}

\item{quantile}{size of quantile which provides value-range of classification}

\item{no_classes}{number of classes to be generated}
}
\value{
dataframe with 4 columns: class (name of each class, here automatically set from A-Z),
        variable (variable chosen for classification), greaterthan (values defining the threshold of each class),
        and color (defining the coloring for the map)
}
\description{
Create initial dataframe for 1-variable classification, to be displayed on the UI table
}
\examples{
df <- create_df_input(
      variable_name = "crops_pc",
      q_0025 = 5,
      q_0975 = 346,
      quantile = 75,
      no_classes = 4
      )

}
