% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plots.R
\name{create_plotly_barplot}
\alias{create_plotly_barplot}
\title{Create interactive stacked barplots of class-distribution for region and axis}
\usage{
create_plotly_barplot(data)
}
\arguments{
\item{data}{classified network data with entries for regional and axis dgos}

\item{colors}{color vector}
}
\value{
interactive stacked barplot with class distribution in % for each scale-group (region and axis)
}
\description{
Create interactive stacked barplots of class-distribution for region and axis
}
\examples{
create_plotly_barplot(data_plots)
}
