% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plots.R
\name{create_plotly_violinplot}
\alias{create_plotly_violinplot}
\title{Create interactive violinplots for a specific variable for region and axis}
\usage{
create_plotly_violinplot(data, var)
}
\arguments{
\item{data}{classified network data with entries for regional and axis dgos}

\item{var}{variable based on which the violinplots should be created}
}
\value{
plotly interactive violinplots for each scale-group (region and axis)
}
\description{
Create interactive violinplots for a specific variable for region and axis
}
\examples{
violinplot_plotly <- create_plotly_violinplot(data_plots, "forest_pc")
}
