% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_fct.R
\name{data_basemaps_df}
\alias{data_basemaps_df}
\title{Get Basemaps Data Frame}
\usage{
data_basemaps_df()
}
\value{
A data frame containing information about available basemaps, with columns 'name', 'url', 'layer', and 'attribution'.
}
\description{
This function retrieves information about available basemaps, including their names, URLs, layers, and attribution.
}
\examples{
\dontrun{
  # Example usage:
  basemaps <- data_basemaps_df()
}

}
