% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data.R
\name{data_get_axes}
\alias{data_get_axes}
\title{Get Network Axes Data}
\usage{
data_get_axes(con)
}
\arguments{
\item{con}{Connection to Postgresql database.}
}
\value{
A sf data frame containing the network axes.
}
\description{
This function retrieves data about the network axes available on the server
}
\examples{
con <- db_con()
axis_data <- data_get_axes(con = con)
DBI::dbDisconnect(con)

}
