% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data.R
\name{data_get_axis_dgos}
\alias{data_get_axis_dgos}
\title{Get Network Metrics Data for a Specific Network Axis}
\usage{
data_get_axis_dgos(selected_axis_id, aggregated = FALSE, con)
}
\arguments{
\item{selected_axis_id}{The ID of the selected network axis.}

\item{aggregated}{A boolean indicating whether to retrieve aggregated data (TRUE) or detailed data (FALSE). Default is FALSE.}

\item{con}{Connection to Postgresql database.}
}
\value{
A sf data frame containing information about network metrics for the specified network axis.
}
\description{
This function retrieves data about network metrics for a specific network axis based on its ID.
}
\examples{
con <- db_con()
network_metrics_data <- data_get_axis_dgos(selected_axis_id = 2000796122, con = con, aggregated=TRUE)
network_metrics_data <- data_get_axis_dgos(selected_axis_id = 2000796122, con = con, aggregated=FALSE)
DBI::dbDisconnect(con)

}
