% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data.R
\name{data_get_axis_start_end}
\alias{data_get_axis_start_end}
\title{Get the start and end coordinates of a spatial object's axis}
\usage{
data_get_axis_start_end(selected_axis_id, con)
}
\arguments{
\item{selected_axis_id}{The ID of the selected axis for which to retrieve the start and end coordinates.}

\item{con}{Connection to Postgresql database.}
}
\value{
A data frame with two rows, where the first row contains the start
        coordinates (x and y) and the second row contains the end coordinates (x and y).
}
\description{
This function takes a spatial object with a LINESTRING geometry and returns
a data frame containing the start and end coordinates of the axis.
}
\examples{

df <- data_get_axis_start_end(line_sf)

}
