% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data.R
\name{data_get_carhyce_stations}
\alias{data_get_carhyce_stations}
\title{Get Stations Carhyce}
\usage{
data_get_carhyce_stations(con)
}
\arguments{
\item{con}{Connection to Postgresql database.}
}
\value{
A sf data frame containing the CarHyCe datapoints.
}
\description{
This function retrieves the datapoints of the CarHyCe hydrological stations.
}
\examples{
con <- db_con()
carhyce_stations <- data_get_carhyce_stations(con = con)
DBI::dbDisconnect(con)

}
