% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_fct.R
\name{data_get_dgo_in_region}
\alias{data_get_dgo_in_region}
\title{Get Network DGO Data in a Region}
\usage{
data_get_dgo_in_region(selected_region_id)
}
\arguments{
\item{selected_region_id}{The ID of the selected region.}
}
\value{
A sf data frame containing information about the network axis within the specified region.
}
\description{
This function retrieves data about the network DGO with the metrics within a specified region based on its ID.
}
\examples{
axis_data <- data_get_dgo_in_region(selected_region_id = 11)

}
