% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data.R
\name{data_get_distr_class}
\alias{data_get_distr_class}
\title{Get statistics on classes distributions for different levels (france, basin, region) and a specified class}
\usage{
data_get_distr_class(con, class_name_selected)
}
\arguments{
\item{con}{Connection to Postgresql database.}

\item{class_name}{Name of the proposed class for which stats should be generated}
}
\value{
a dataframe which contains the statistics for the specified class for different entities: France, Basins, Regions available on the serve
}
\description{
Get statistics on classes distributions for different levels (france, basin, region) and a specified class
}
\examples{
con=db_con()
data_get_distr_class(con,"class_style")
}
