% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data.R
\name{data_get_distr_class_man}
\alias{data_get_distr_class_man}
\title{Get statistics on classes distributions for different levels (france, basin, region) and manually specified classes}
\usage{
data_get_distr_class_man(con, manual_classes_table)
}
\arguments{
\item{con}{Connection to Postgresql database.}

\item{manual_classes_table}{Dataframe of manually defined classes, with 4 columns: variable, class, greaterthan, color}
}
\value{
Dataframe which contains the statistics for the specified class for different entities: France, Basins, Regions available on the server
}
\description{
Get statistics on classes distributions for different levels (france, basin, region) and manually specified classes
}
