% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data.R
\name{data_get_elevation_profiles}
\alias{data_get_elevation_profiles}
\title{Get elevation profiles data from selected dgo fid.}
\usage{
data_get_elevation_profiles(selected_dgo_fid)
}
\arguments{
\item{selected_dgo_fid}{integer selected dgo fid}
}
\value{
data.frame
}
\description{
Get elevation profiles data from selected dgo fid.
}
\examples{
data_get_elevation_profiles(selected_dgo_fid = 29567)
}
