% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data.R
\name{data_get_hydro_sites}
\alias{data_get_hydro_sites}
\title{Get hydrometric sites.}
\usage{
data_get_hydro_sites(con)
}
\arguments{
\item{con}{Connection to Postgresql database.}
}
\value{
sf data frame containing the datapoints of the hydrometric sites available on the server
}
\description{
This function retrieves the locations of the hydrometric sites from Hubeau.
}
\examples{
con <- db_con()
hydro_sites <- data_get_hydro_sites(con = con)
DBI::dbDisconnect(con)

}
