% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data.R
\name{data_get_levels_names}
\alias{data_get_levels_names}
\title{Get the names of the hydrographic basins and regions}
\usage{
data_get_levels_names(con)
}
\arguments{
\item{con}{Connection to Postgresql database.}
}
\description{
Get the names of the hydrographic basins and regions
}
