% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data.R
\name{data_get_metrics}
\alias{data_get_metrics}
\title{This function retrieves the aggregated metrics for the whole network.}
\usage{
data_get_metrics(con)
}
\arguments{
\item{con}{Connection to Postgresql database.}
}
\value{
sf data frame containing the datapoints of the homogenous segments' metrics
}
\description{
This function retrieves the aggregated metrics for the whole network.
}
\examples{
con <- db_con()
hydro_sites <- data_get_metrics(con = con)
DBI::dbDisconnect(con)

}
