% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_fct.R
\name{data_get_min_max_strahler}
\alias{data_get_min_max_strahler}
\title{Get Minimum and Maximum Strahler Values for a Selected Region}
\usage{
data_get_min_max_strahler(selected_region_id = region_click_id())
}
\arguments{
\item{selected_region_id}{The ID of the selected region.}
}
\value{
A data frame containing two columns: 'min' and 'max', representing the minimum and maximum Strahler values for the specified region.
}
\description{
This function retrieves the minimum and maximum values of the Strahler metric for a specified region.
}
\examples{
\dontrun{
  # Example usage:
  data <- data_get_min_max_strahler(selected_region_id = 1)
}

}
