% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data.R
\name{data_get_network_axis}
\alias{data_get_network_axis}
\title{Get Network Metrics Data for a Specific Network Axis}
\usage{
data_get_network_axis(selected_axis_id)
}
\arguments{
\item{selected_axis_id}{The ID of the selected network axis.}
}
\value{
A sf data frame containing information about network metrics for the specified network axis.
}
\description{
This function retrieves data about network metrics for a specific network axis based on its ID.
}
\examples{
network_metrics_data <- data_get_network_axis(selected_axis_id = 2000796122)

}
