% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data.R
\name{data_get_network_region}
\alias{data_get_network_region}
\title{Get all Network Metrics Data for a Specific region}
\usage{
data_get_network_region(selected_region_id, con)
}
\arguments{
\item{selected_region_id}{The ID of the selected region}

\item{con}{PqConnection to Postgresql database.}
}
\value{
A sf data frame containing information about network metrics for the specified region
}
\description{
This function retrieves data about network metrics for a specific region based on its ID.
}
\examples{
con <- db_con()
network_metrics_data <- data_get_network_axis(selected_region_id = 11, con = con)
DBI::dbDisconnect(con)

}
