% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_fct.R
\name{data_get_region}
\alias{data_get_region}
\title{Get hydrological region selected by user}
\usage{
data_get_region(region_click_id)
}
\arguments{
\item{region_click_id}{The ID of the selected region.}
}
\value{
A sf data frame containing hydrographical data for the specified region.
}
\description{
This function retrieves hydrographical data for a specified region based on its ID.
}
\examples{
data <- data_get_region(region_click_id = 11)

}
