% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data.R
\name{data_get_regions}
\alias{data_get_regions}
\title{Get hydrological regions}
\usage{
data_get_regions(con, opacity = list(clickable = 0.01, not_clickable = 0.1))
}
\arguments{
\item{con}{Connection to Postgresql database.}
}
\value{
A sf data frame containing hydrological regions.
}
\description{
This function retrieves all hydrological regions available on the server and converts them to sf
}
\examples{
con <- db_con()
data <- data_get_regions(con = con)
DBI::dbDisconnect(con)

}
