% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_fct.R
\name{data_get_roe_in_region}
\alias{data_get_roe_in_region}
\title{Get Referentiel des Obstacles aux Ecoulement Data for a Region}
\usage{
data_get_roe_in_region(selected_region_id)
}
\arguments{
\item{selected_region_id}{The ID of the selected region.}
}
\value{
A sf data frame containing information about ROE within the specified region.
}
\description{
This function retrieves data about Referentiel des Obstacles aux Ecoulement (ROE) within a specified region based on its ID.
}
\examples{
roe_data <- data_get_roe_in_region(selected_region_id = 11)

}
