% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data.R
\name{data_get_roe_sites}
\alias{data_get_roe_sites}
\title{Get Referentiel des Obstacles aux Ecoulement Data}
\usage{
data_get_roe_sites(con)
}
\arguments{
\item{con}{Connection to Postgresql database.}
}
\value{
A sf data frame containing the ROE datapoints.
}
\description{
This function retrieves the datapoints of the Referentiel des Obstacles aux Ecoulement (ROE).
}
\examples{
con <- db_con()
roe_data <- data_get_roe_sites(con = con)
DBI::dbDisconnect(con)

}
