% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data.R
\name{data_get_stats_metrics}
\alias{data_get_stats_metrics}
\title{Get statistics on network metrics for different levels (france, basin, region)}
\usage{
data_get_stats_metrics(con)
}
\arguments{
\item{con}{Connection to Postgresql database.}
}
\value{
Dataframe which contains the statistics for all metrics for different entities: France, Basins, Regions
}
\description{
Get statistics on network metrics for different levels (france, basin, region)
}
