% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_fct.R
\name{data_overlayers_df}
\alias{data_overlayers_df}
\title{Get Overlayers Data Frame}
\usage{
data_overlayers_df()
}
\value{
A data frame containing information about available overlayers, with columns 'name', 'url', 'layer', and 'attribution'.
}
\description{
This function retrieves information about available overlayers, including their names, URLs, layers, and attribution.
}
\examples{
\dontrun{
  # Example usage:
  overlayers <- data_overlayers_df()
}

}
