% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_table_characteristics.R
\name{fct_table_create_reactable}
\alias{fct_table_create_reactable}
\title{create a table to compare metric-characteristics of region, axis, and segment}
\usage{
fct_table_create_reactable(df, unit)
}
\arguments{
\item{df}{dataframe of merged regional, axis and dgo data}

\item{unit}{}
}
\value{
reactable table object
fct_table_create_table_df(region_sf = network_dgo,
 axis_sf = network_dgo,
 dgo_sf = NULL)
}
\description{
create a table to compare metric-characteristics of region, axis, and segment
}
