% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_table_characteristics.R
\name{fct_table_create_table_df}
\alias{fct_table_create_table_df}
\title{Merge regional, axis, and dgo dataframes together for reactable table}
\usage{
fct_table_create_table_df(region_pivot, axis_pivot, dgo_pivot)
}
\arguments{
\item{region_pivot}{pivoted df from selected regional network}

\item{axis_pivot}{pivoted df from selected axis network}

\item{dgo_pivot}{pivoted df from selected dgo}
}
\value{
dataframe with metrics as rows
}
\description{
Merge regional, axis, and dgo dataframes together for reactable table
}
\examples{
fct_table_create_table_df(network_dgo, network_dgo, network_dgo[1,])
}
