% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_table_characteristics.R
\name{fct_table_pivot_sf}
\alias{fct_table_pivot_sf}
\title{convert sf to pivoted dataframe with variables as rows}
\usage{
fct_table_pivot_sf(df_sf)
}
\arguments{
\item{df_sf}{sf object with one or several dgos}
}
\description{
convert sf to pivoted dataframe with variables as rows
}
