% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_fct.R
\name{get_axis}
\alias{get_axis}
\title{get axis network data}
\usage{
get_axis(selected_region_id = region_click$id)
}
\arguments{
\item{selected_region_id}{selected region id by user}
}
\value{
sf data.frame
}
\description{
get axis network data
}
\examples{
get_axis(selected_region_id = click_value()$id)
}
