% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_fct.R
\name{get_bassins}
\alias{get_bassins}
\title{get hydrographic bassin from database}
\usage{
get_bassins()
}
\value{
sf data.frame
}
\description{
get hydrographic bassin from database
}
\examples{
bassin_hydro <- get_bassins()
}
