% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plots.R
\name{get_colors_char_df}
\alias{get_colors_char_df}
\title{create dataframe of color-classes and values}
\usage{
get_colors_char_df(data)
}
\arguments{
\item{data}{classified network with corresponding colors for each class}
}
\value{
color vector
}
\description{
create dataframe of color-classes and values
}
\examples{
get_colors_char_df(network)
}
