% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_fct.R
\name{get_network_region_with_metrics}
\alias{get_network_region_with_metrics}
\title{get network in region selected with metrics data}
\usage{
get_network_region_with_metrics(selected_region_id = region_click$id)
}
\arguments{
\item{selected_region_id}{}
}
\value{
sf data.frame
}
\description{
get network in region selected with metrics data
}
\examples{
data <- get_network_region_with_metrics(selected_region_id = region_click$id)
}
