% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_fct.R
\name{get_region}
\alias{get_region}
\title{get region clicked by user}
\usage{
get_region(region_click_id = region_click$id)
}
\arguments{
\item{region_click_id}{}
}
\value{
sf data.frame
}
\description{
get region clicked by user
}
\examples{
selected_region_feature <- get_region(region_click_id = region_click$id)
}
