% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_fct.R
\name{get_regions_in_bassin}
\alias{get_regions_in_bassin}
\title{get the hydrographic regions from the selected bassin}
\usage{
get_regions_in_bassin(selected_bassin_id = bassin_click$id)
}
\arguments{
\item{selected_region}{}
}
\value{
sf data.frame
}
\description{
get the hydrographic regions from the selected bassin
}
\examples{
region_hydro <- get_regions_in_bassin(selected_bassin_id = bassin_click$id)
}
