% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_fct.R
\name{get_roe_in_region}
\alias{get_roe_in_region}
\title{get ROE data from PostgreSQL}
\usage{
get_roe_in_region(selected_region_id = region_click$id)
}
\arguments{
\item{selected_region_id}{selected region id by user}
}
\value{
sf data.frame
}
\description{
get ROE data from PostgreSQL
}
\examples{
get_roe_in_region(region_click$id)
}
