% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot_long.R
\name{lg_add_trace}
\alias{lg_add_trace}
\title{plotly add trace.}
\usage{
lg_add_trace(data, y, y_label, yaxis = "y1", color = "#22223b")
}
\arguments{
\item{data}{data frame containing the selected axis data.}

\item{y}{text metric to be plotted on the y-axis.}

\item{y_label}{text name of the metric plotted.}

\item{yaxis}{text axis id.}
}
\value{
list
}
\description{
plotly add trace.
}
