% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_lg_profile.R
\name{lg_annotations_layout}
\alias{lg_annotations_layout}
\title{plotly annotations layout.}
\usage{
lg_annotations_layout(data)
}
\arguments{
\item{data}{data.frame dgo from axis.}
}
\value{
list
}
\description{
plotly annotations layout.
}
