% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot_long.R
\name{lg_profile_empty}
\alias{lg_profile_empty}
\title{Create an empty longitudinal profile plot.}
\usage{
lg_profile_empty()
}
\value{
An empty longitudinal profile plot with a specified title.
}
\description{
This function generates an empty longitudinal profile plot using the 'plot_ly'
function from the 'plotly' package.
}
\examples{
# Create an empty longitudinal profile plot
empty_plot <- lg_profile_empty()
empty_plot

}
