% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot_long.R
\name{lg_profile_first}
\alias{lg_profile_first}
\title{Add first metric for longitudinal profile plot for selected axis data}
\usage{
lg_profile_first(data, y, y_label, y_label_category)
}
\arguments{
\item{data}{A data frame containing the selected axis data.}

\item{y}{The primary metric to be plotted on the left y-axis.}

\item{y_label}{The name of the metric plotted.}

\item{y_label_category}{The metric category name.}
}
\value{
A dual-axis longitudinal profile plot with the specified metrics.
}
\description{
This function adds the first metric trace to the plot
}
\examples{
\dontrun{
# like lg_profile_update_main function, see example in documentation
}

}
