% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lg_profile_fct.R
\name{lg_profile_second}
\alias{lg_profile_second}
\title{Create a dual-axis longitudinal profile plot for selected axis data}
\usage{
lg_profile_second(
  data = selected_axis_df,
  y = "active_channel_width",
  y2 = "talweg_elevation_min"
)
}
\arguments{
\item{data}{A data frame containing the selected axis data.}

\item{y}{The primary metric to be plotted on the left y-axis.}

\item{y2}{The secondary metric to be plotted on the right y-axis.}
}
\value{
A dual-axis longitudinal profile plot with the specified metrics.
}
\description{
This function generates a dual-axis longitudinal profile plot using the 'plot_ly'
function from the 'plotly' package. It allows you to visualize two different
metrics along the selected axis.
}
\examples{
# Create a dual-axis longitudinal profile plot for active channel width and
# talweg elevation min
dual_axis_plot <- lg_profile_second(data = selected_axis_df,
                                   y = "active_channel_width",
                                   y2 = "talweg_elevation_min")

}
