% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot_long.R
\name{lg_roe_vertical_line}
\alias{lg_roe_vertical_line}
\title{Create the ROE vertical lines for plotly shapes}
\usage{
lg_roe_vertical_line(roe_distance_axis)
}
\arguments{
\item{roe_distance_axis}{vector ROE distance on axis.}
}
\value{
list of each vertical lines
}
\description{
Create the ROE vertical lines for plotly shapes
}
