% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lg_profile_fct.R
\name{lg_vertical_line}
\alias{lg_vertical_line}
\title{Create a vertical dashed line annotation for longitudinal profile plots}
\usage{
lg_vertical_line(x = 0, color = "green")
}
\arguments{
\item{x}{The x-coordinate where the vertical line should be positioned.}

\item{color}{The color of the vertical dashed line (default is "green").}
}
\value{
A list object representing a vertical dashed line annotation.
}
\description{
This function generates a vertical dashed line annotation for longitudinal profile
plots using the 'plotly' package.
}
\examples{
# Create a vertical dashed line annotation at x = 10 with a red color
vertical_line_annotation <- lg_vertical_line(x = 10, color = "red")

}
