% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_lg_profile.R
\name{lg_xaxis_layout}
\alias{lg_xaxis_layout}
\title{plotly xaxis layout.}
\usage{
lg_xaxis_layout(data)
}
\arguments{
\item{data}{data.frame dgo from axis.}
}
\value{
list
}
\description{
plotly xaxis layout.
}
