% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot_long.R
\name{lg_yaxis_layout}
\alias{lg_yaxis_layout}
\title{plotly yaxis layout.}
\usage{
lg_yaxis_layout(y_label_category, y_label)
}
\arguments{
\item{y_label_category}{text metric category name.}

\item{y_label}{text name of the metric plotted.}
}
\value{
list
}
\description{
plotly yaxis layout.
}
