% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_map.R
\name{map_add_axis_dgos}
\alias{map_add_axis_dgos}
\title{Add Axis Data to an Existing Leaflet Map}
\usage{
map_add_axis_dgos(map, axis_data, Proposed_class, group)
}
\arguments{
\item{map}{An existing Leaflet map to which axis data will be added.}

\item{data_axis}{A sf data frame containing axis data.}
}
\value{
An updated Leaflet map with axis data added.
}
\description{
This function adds axis data as polylines to an existing Leaflet map. The axis are transparent,
but when hovering over they appear red.
}
