% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_fct.R
\name{map_add_basemaps}
\alias{map_add_basemaps}
\title{Add all the basemaps}
\usage{
map_add_basemaps(map, basemaps)
}
\arguments{
\item{map}{leaflet map}

\item{basemaps}{basemap data.frame}
}
\value{
basemaps for leaflet
}
\description{
Add all the basemaps
}
\examples{
map \%>\%
map_add_basemaps(data_basemaps_df())
}
