% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_map.R
\name{map_add_network}
\alias{map_add_network}
\title{Map the hydrographique network}
\usage{
map_add_network(
  map,
  wms_params_network,
  group,
  cql_filter = "",
  style = "mapdo:classes_proposed_strahler"
)
}
\arguments{
\item{map}{An existing Leaflet map to which WMS tiles will be added.}

\item{cql_filter}{A CQL filter to apply to the WMS request.}

\item{wms_params}{list containing the specific network-WMS parameters.}
}
\value{
An updated Leaflet map with WMS tiles of the styled network. Standard styling
according to strahler-order and standard regional constraint: whole France.
}
\description{
Adds the hydrographique network with a custom styling and regional constraint to the map
}
\examples{
map \%>\%
map_add_network(wms_params = params_wms()$network, cql_filter = "gid_region <> 11", style = "mapdo:classes_proposed_urban")

}
