% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_fct.R
\name{map_add_overlayers}
\alias{map_add_overlayers}
\title{Add all the overlayers}
\usage{
map_add_overlayers(map, overlayers)
}
\arguments{
\item{map}{leaflet map}

\item{overlayers}{overlayers data.frame}
}
\value{
overlayers for leaflet
}
\description{
Add all the overlayers
}
\examples{
map \%>\%
map_add_overlayers(data_overlayers_df())
}
