% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_fct.R
\name{map_add_regions_in_bassin}
\alias{map_add_regions_in_bassin}
\title{Update initial map to hydrographic regions}
\usage{
map_add_regions_in_bassin(
  map,
  bassin_click = bassin_click,
  regions_data = region_hydro,
  bassins_group = "BASSIN",
  regions_group = "REGION"
)
}
\arguments{
\item{map}{map object for pipe}

\item{bassin_click}{bassin selected by user}

\item{regions_data}{hydrographic regions data}

\item{bassins_group}{bassins layer group}

\item{regions_group}{regions layer group}
}
\value{
leaflet updated
}
\description{
Update initial map to hydrographic regions
}
\examples{
map_add_regions_in_bassin(bassin_click = bassin_click,
regions_data = region_hydro,
bassins_group = "BASSIN",
regions_group = "REGION")
}
