% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_map.R
\name{map_add_wms_overlayers}
\alias{map_add_wms_overlayers}
\title{Add Overlayer Layers to an Existing Leaflet Map}
\usage{
map_add_wms_overlayers(map)
}
\arguments{
\item{map}{An existing Leaflet map to which overlayer layers will be added.}
}
\value{
An updated Leaflet map with overlayer layers added.
}
\description{
This function adds overlayer layers to an existing Leaflet map.
}
\examples{
\dontrun{
  # Used in map_region_clicked() function, its use is in the function
}

}
