% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_map.R
\name{map_axis}
\alias{map_axis}
\title{Add Axis Data to an Existing Leaflet Map}
\usage{
map_axis(map, data_axis)
}
\arguments{
\item{map}{An existing Leaflet map to which axis data will be added.}

\item{data_axis}{A sf data frame containing axis data.}
}
\value{
An updated Leaflet map with axis data added.
}
\description{
This function adds axis data as polylines to an existing Leaflet map.
}
\examples{
\dontrun{
  # Used in map_metric() function, see full example in map_metric() documentation
}

}
