% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_fct.R
\name{map_axis_no_metric}
\alias{map_axis_no_metric}
\title{Update map with axis without metric selected}
\usage{
map_axis_no_metric(map, data_axis = network_region_axis(), axis_group = "AXIS")
}
\arguments{
\item{map}{map to update}

\item{data_axis}{axis data}

\item{axis_group}{axis layer group}
}
\value{
update leaflet map
}
\description{
Update map with axis without metric selected
}
\examples{
leafletProxy("exploremap") \%>\%
  map_axis_no_metric(data_axis = network_region_axis(), axis_group = "AXIS")
}
