% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_fct.R
\name{map_axis_start_end}
\alias{map_axis_start_end}
\title{Add start and end markers to a leaflet map}
\usage{
map_axis_start_end(map, axis_start_end = axis_start_end())
}
\arguments{
\item{map}{A Leaflet map object created using the 'leaflet' package.}

\item{axis_start_end}{A data frame containing start and end coordinates with
columns 'X' for longitude and 'Y' for latitude.}
}
\value{
A Leaflet map object with start and end markers added.
}
\description{
This function adds start and end markers to a Leaflet map based on the provided
start and end coordinates.
}
\examples{
# Create a simple Leaflet map
my_map <- leaflet() \%>\%
  setView(lng = -73.985, lat = 40.748, zoom = 12)

# Create a data frame with start and end coordinates
coordinates_df <- data.frame(
  X = c(-73.985, -73.995),
  Y = c(40.748, 40.755)
)

# Add start and end markers to the map
my_map <- map_axis_start_end(my_map, coordinates_df)
my_map

}
