% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_fct.R
\name{map_dgo_axis}
\alias{map_dgo_axis}
\title{Add DGO axis to a Leaflet map}
\usage{
map_dgo_axis(map, selected_axis, region_axis)
}
\arguments{
\item{map}{A Leaflet map object.}

\item{selected_axis}{A data frame containing selected axe to be displayed.}

\item{region_axis}{A data frame containing region-specific axes to be displayed.}
}
\value{
A modified Leaflet map object with DGO axes added.
}
\description{
This function adds DGO axis to a Leaflet map with the option to highlight selected axes.
}
\examples{
# Create a basic Leaflet map
my_map <- leaflet() \%>\%
  setView(lng = -73.985, lat = 40.748, zoom = 12)

# Define selected and region-specific axes data frames
selected_axes <- data.frame(...)
region_axes <- data.frame(...)

# Add DGO axes to the map
my_map <- map_dgo_axis(my_map, selected_axes, region_axes)
my_map

}
