% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_map.R
\name{map_dgo_cross_section}
\alias{map_dgo_cross_section}
\title{Highlight DGO clicked for cross section.}
\usage{
map_dgo_cross_section(map, selected_dgo)
}
\arguments{
\item{map}{A Leaflet map object.}

\item{selected_dgo}{sf dgo clicked.}
}
\value{
Leaflet map
}
\description{
Highlight DGO clicked for cross section.
}
