% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_fct.R
\name{map_init_bassins}
\alias{map_init_bassins}
\title{Map initialization with hydrographic bassins data}
\usage{
map_init_bassins(bassins_data = get_bassins(), group = "A")
}
\arguments{
\item{bassins_data}{hydrographic bassins sf}

\item{group}{layer group}
}
\value{
leaflet map
}
\description{
Map initialization with hydrographic bassins data
}
\examples{
map <- map_init_bassins(bassins_data = get_bassins(), group = "A")
}
