% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_map.R
\name{map_legend_metric}
\alias{map_legend_metric}
\title{Generate and display a legend for a map layer using a provided SLD (Styled Layer Descriptor) body.}
\usage{
map_legend_metric(sld_body, wms_params)
}
\arguments{
\item{sld_body}{A character string containing the SLD (Styled Layer Descriptor) body that defines the map layer's styling.}

\item{wms_params}{list containing the specific network-WMS parameters.}
}
\value{
An HTML img tag representing the legend for the specified map layer.
}
\description{
This function constructs a legend for a map layer by sending a GetLegendGraphic request to a WMS (Web Map Service) server.
}
