% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_fct.R
\name{map_legend_vector_overlayer}
\alias{map_legend_vector_overlayer}
\title{Generate a legend entry for a vector overlay layer.}
\usage{
map_legend_vector_overlayer(layer_label)
}
\arguments{
\item{layer_label}{A character string representing the label or name of the vector overlay layer.}
}
\value{
An HTML div element representing the legend entry for the vector overlay layer.
}
\description{
This function generates an HTML representation of a legend entry for a vector overlay layer. The legend entry consists of a colored circle with a label indicating the layer's name.
}
\examples{
# Create a legend entry for a vector overlay layer
legend_entry <- map_legend_vector_overlayer(layer_label = "ROE")
print(legend_entry)

}
