% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_map.R
\name{map_legend_wms_overlayer}
\alias{map_legend_wms_overlayer}
\title{Generate and display a legend for a WMS (Web Map Service) layer overlay using specified WMS parameters.}
\usage{
map_legend_wms_overlayer(wms_params)
}
\arguments{
\item{wms_params}{A list containing the following parameters for the WMS layer overlay:
\itemize{
  \item \code{language} (character): The language to use for the legend.
  \item \code{version} (character): The version of the WMS service.
  \item \code{service} (character): The WMS service type (e.g., "WMS").
  \item \code{sld_version} (character): The SLD (Styled Layer Descriptor) version.
  \item \code{layer} (character): The name of the WMS layer for which the legend is generated.
  \item \code{format} (character): The desired format of the legend image (e.g., "image/png").
  \item \code{style} (character): The style to use for rendering the legend.
  \item \code{url} (character): The URL of the WMS server.
}}
}
\value{
An HTML div element containing an img tag representing the legend for the specified WMS layer overlay.
}
\description{
This function constructs a legend for a WMS layer overlay by sending a GetLegendGraphic request to a WMS server.
}
\examples{
# Define WMS parameters for a layer overlay like
wms_params <- params_wms()$inondation

# Generate and display the legend for the WMS layer overlay
legend <- map_legend_wms_overlayer(wms_params)
print(legend)

}
