% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_fct.R
\name{map_metric}
\alias{map_metric}
\title{Add a metric layer with custom symbology to a map.}
\usage{
map_metric(
  map,
  wms_params = params_wms()$metric,
  cql_filter = "",
  sld_body = "",
  data_axis = network_region_axis()
)
}
\arguments{
\item{map}{A leaflet map object to which the metric layer will be added.}

\item{wms_params}{A list containing WMS parameters for the metric layer. If not provided, default parameters are retrieved using the \code{\link{params_wms}} function.}

\item{cql_filter}{A character string representing a CQL filter to apply to the metric layer.}

\item{sld_body}{A character string representing the SLD (Styled Layer Descriptor) body for custom styling of the metric layer.}

\item{data_axis}{A data axis to display on the map.}
}
\value{
A leaflet map object with the metric layer added.
}
\description{
This function adds a metric layer with custom symbology to a leaflet map. It allows you to specify custom parameters for the Web Map Service (WMS) request, apply a CQL (Common Query Language) filter, and provide a custom SLD (Styled Layer Descriptor) body for styling the layer. Additionally, you can specify the data axis to display on the map.
}
\examples{
# Create a leaflet map
my_map <- leaflet() \%>\%
  addTiles() \%>\%
  setView(lng = 0, lat = 0, zoom = 3)

# Add a custom metric layer to the map
map_metric(my_map, wms_params = params_wms()$metric, cql_filter = "value > 100", sld_body = my_custom_sld, data_axis = my_axis_data)

}
