% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_fct.R
\name{map_network_no_metric}
\alias{map_network_no_metric}
\title{Update map with network without metric selected}
\usage{
map_network_no_metric(map, datamap = network_filter(), network_group = "D")
}
\arguments{
\item{map}{map shiny id}

\item{datamap}{network data}

\item{network_group}{network group}
}
\value{
leaflet update
}
\description{
Update map with network without metric selected
}
\examples{
map_metric("exploremap", network_filter(), varsel(), network_group = "D")
}
