% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_fct.R
\name{map_region_clicked}
\alias{map_region_clicked}
\title{Update Leaflet Map for a Clicked Region}
\usage{
map_region_clicked(
  map,
  region_click = region_click,
  selected_region_feature = selected_region_feature
)
}
\arguments{
\item{map}{An existing Leaflet map to be updated.}

\item{region_click}{A vector containing information about the clicked region.}

\item{selected_region_feature}{A sf data frame containing information about the selected region feature.}
}
\value{
An updated Leaflet map with relevant layers and information displayed.
}
\description{
This function updates an existing Leaflet map when a region is clicked, displaying the region and overlayers.
}
\examples{
\dontrun{
  # Example usage:
  updated_map <- map_region_clicked(map = existing_map,
    region_click = clicked_region_data, selected_region_feature = selected_feature_data)
}

}
