% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_fct.R
\name{map_wms_metric}
\alias{map_wms_metric}
\title{Add WMS Tiles with Metric Data to an Existing Leaflet Map}
\usage{
map_wms_metric(
  map,
  wms_params = params_wms()$metric,
  cql_filter = "",
  sld_body = ""
)
}
\arguments{
\item{map}{An existing Leaflet map to which WMS tiles will be added.}

\item{wms_params}{A list of WMS parameters.}

\item{cql_filter}{A CQL filter to apply to the WMS request.}

\item{sld_body}{A custom SLD (Styled Layer Descriptor) body for symbology customization.}
}
\value{
An updated Leaflet map with WMS tiles containing metric data added.
}
\description{
This function adds WMS tiles with metric data to an existing Leaflet map, allowing for customization of style and filtering.
}
\examples{
\dontrun{
  # Example usage:
  updated_map <- map_wms_metric(map = existing_map, wms_params = params_wms()$metric, cql_filter = "metric > 100", sld_body = "<sld>...</sld>")
}

}
