% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_analysis.R
\name{merge_regional_axis_dfs}
\alias{merge_regional_axis_dfs}
\title{Combine classified regional network and axis network in one frame}
\usage{
merge_regional_axis_dfs(data_region, data_axis, var)
}
\arguments{
\item{data_region}{sf-df with all dgos of selected region}

\item{data_axis}{sf-df with all dgos of selected axis}

\item{var}{variable for which classification was undertaken}
}
\value{
merged df with regional and axis dgos, identifiable by factor-variable "scale"
}
\description{
Combine classified regional network and axis network in one frame
}
\examples{
merge_regional_axis_dfs(data_classified,
                        data_classified \%>\% filter(toponyme == "l'Isère"),
                        "forest_pc")
}
