% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_explore.R
\name{mod_explore_ui}
\alias{mod_explore_ui}
\title{explore UI Function}
\usage{
mod_explore_ui(id)
}
\arguments{
\item{id}{shiny id}

\item{input, output, session}{Internal parameters for {shiny}.}
}
\description{
A shiny Module.
}
