% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_network_axis.R
\docType{data}
\name{network_axis}
\alias{network_axis}
\title{network_axis}
\format{
A data frame with 3 rows and 4 variables:
\describe{
  \item{ fid }{  integer }
  \item{ axis }{  numeric }
  \item{ gid_region }{  integer }
  \item{ geom }{  sfc_MULTILINESTRING,sfc }
}
}
\source{
Source
}
\usage{
network_axis
}
\description{
Description.
}
\keyword{datasets}
