% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_network_dgo.R
\docType{data}
\name{network_dgo}
\alias{network_dgo}
\title{network_dgo}
\format{
A data frame with 573 rows and 46 variables:
\describe{
  \item{ fid }{  integer }
  \item{ axis }{  numeric }
  \item{ measure }{  integer }
  \item{ toponyme }{  character }
  \item{ strahler }{  integer }
  \item{ talweg_elevation_min }{  numeric }
  \item{ active_channel_width }{  numeric }
  \item{ natural_corridor_width }{  numeric }
  \item{ connected_corridor_width }{  numeric }
  \item{ valley_bottom_width }{  numeric }
  \item{ talweg_slope }{  numeric }
  \item{ floodplain_slope }{  numeric }
  \item{ water_channel }{  numeric }
  \item{ gravel_bars }{  numeric }
  \item{ natural_open }{  numeric }
  \item{ forest }{  numeric }
  \item{ grassland }{  numeric }
  \item{ crops }{  numeric }
  \item{ diffuse_urban }{  numeric }
  \item{ dense_urban }{  numeric }
  \item{ infrastructures }{  numeric }
  \item{ active_channel }{  numeric }
  \item{ riparian_corridor }{  numeric }
  \item{ semi_natural }{  numeric }
  \item{ reversible }{  numeric }
  \item{ disconnected }{  numeric }
  \item{ built_environment }{  numeric }
  \item{ water_channel_pc }{  numeric }
  \item{ gravel_bars_pc }{  numeric }
  \item{ natural_open_pc }{  numeric }
  \item{ forest_pc }{  numeric }
  \item{ grassland_pc }{  numeric }
  \item{ crops_pc }{  numeric }
  \item{ diffuse_urban_pc }{  numeric }
  \item{ dense_urban_pc }{  numeric }
  \item{ infrastructures_pc }{  numeric }
  \item{ active_channel_pc }{  numeric }
  \item{ riparian_corridor_pc }{  numeric }
  \item{ semi_natural_pc }{  numeric }
  \item{ reversible_pc }{  numeric }
  \item{ disconnected_pc }{  numeric }
  \item{ built_environment_pc }{  numeric }
  \item{ sum_area }{  numeric }
  \item{ idx_confinement }{  numeric }
  \item{ gid_region }{  integer }
  \item{ geom }{  sfc_MULTILINESTRING,sfc }
}
}
\source{
Source
}
\usage{
network_dgo
}
\description{
Description.
}
\keyword{datasets}
