% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_params.R
\name{params_classes}
\alias{params_classes}
\title{Get classes names and description for fluvial styles}
\usage{
params_classes()
}
\value{
tibble with names of classes and their descriptions
}
\description{
Get classes names and description for fluvial styles
}
