% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_params.R
\name{params_classes_colors}
\alias{params_classes_colors}
\title{Get classes names and colors for proposed classifications}
\usage{
params_classes_colors()
}
\value{
list with classes names and colors, identifiable by each classification-name
}
\description{
Get classes names and colors for proposed classifications
}
\examples{
params_classes_colors()$class_habitat
}
