% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_analysis.R
\name{params_get_metric_choices}
\alias{params_get_metric_choices}
\title{get nested list-object with all variables for Metric-selection in selectInput()-Elements}
\usage{
params_get_metric_choices()
}
\value{
list-object with first level the names of metric types and second levels the corresponding metrics for each type
}
\description{
get nested list-object with all variables for Metric-selection in selectInput()-Elements
}
\examples{
params_get_metric_choices()
}
