% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_params.R
\name{params_hubeau_url}
\alias{params_hubeau_url}
\title{create Hubeau url API from map bounds}
\usage{
params_hubeau_url(leafletmap_bounds)
}
\arguments{
\item{leafletmap_bounds}{list leaflet map bounds event.}
}
\value{
character
}
\description{
create Hubeau url API from map bounds
}
\examples{
url <- params_hubeau_url(list(north = 52.2950422845374,
                              east = 10.546875,
                              south = 40.2124407182865,
                              west = -5.625))
}
