% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params_fct.R
\name{params_metrics_choice}
\alias{params_metrics_choice}
\title{Get Choices for Network Metrics}
\usage{
params_metrics_choice()
}
\value{
A list containing choices for various network metrics categories and their corresponding metrics.
}
\description{
This function retrieves a list of choices for selecting network metrics.
}
\examples{
\dontrun{
  # Example usage:
  metrics_choices <- params_metrics_choice()
}

}
