% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_analysis.R
\name{params_metrics_choice_analysis}
\alias{params_metrics_choice_analysis}
\title{Get Choices for Metric Selection}
\usage{
params_metrics_choice_analysis()
}
\value{
A list of choices for selecting metric type."
}
\description{
This function returns a list of choices for selecting metrics organized into categories.
}
\examples{
metric_choices <- params_metrics_choice_analysis()

}
