% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_params.R
\name{params_unit_area}
\alias{params_unit_area}
\title{Get a list of available unit areas.}
\usage{
params_unit_area()
}
\value{
A named character vector containing unit areas and their labels.
}
\description{
This function returns a vector of available unit areas with their respective labels.
}
\examples{
unit_areas <- params_unit_area()
unit_areas["Hectares"]

}
