% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_params.R
\name{params_wms}
\alias{params_wms}
\title{Define Web Map Service (WMS) parameters for different map layers and basemaps.}
\usage{
params_wms()
}
\value{
A list containing WMS parameters for different map layers and basemaps.
}
\description{
This function defines a set of WMS parameters for various map layers and basemaps. The parameters include information such as the layer name, URL of the WMS server, version, format, style, and more. These parameters are organized into a list, making it easy to configure and access them for map display and legend generation.
}
\examples{
# Retrieve WMS parameters for a specific map layer
wms_params <- params_wms()
metric_wms_params <- wms_params$metric

# Access specific WMS parameters
metric_name <- metric_wms_params$name
metric_url <- metric_wms_params$url

}
