% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_analysis.R
\name{plot_class_series_plotly}
\alias{plot_class_series_plotly}
\title{Plot variable series with categorical variable in background}
\usage{
plot_class_series_plotly(data, x, y, cat, colors)
}
\arguments{
\item{data}{input data table}

\item{x}{longitudinal variable}

\item{y}{var used for geom_line()-function}

\item{colors}{vector of colors used for printing category}

\item{var_cat}{categorical var to be plotted in background}
}
\value{
ggplot()-graph of dataseries plotted as continuous line and categorical variable in background
}
\description{
Plot variable series with categorical variable in background
}
\examples{
plot_class_series_plotly(isere, "measure", "PC1", "cluster",
                c("1" = "#bc4749", "2" = "#ffc300", "3" = "#a7c957", "4" = "#1a535c", "5" = "#a9def9"))
}
