% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_analysis_tables.R
\name{prepare_regions_stats_for_table}
\alias{prepare_regions_stats_for_table}
\title{Prepare metrics-statistics dataframe for reactable table}
\usage{
prepare_regions_stats_for_table(data, region_names = NULL)
}
\arguments{
\item{data}{metrics-statistics dataframe}

\item{type}{type of level_type to be included in the dataframe, default is c("Région (total)", "Région")}
}
\value{
metrics-statistics dataframe with *_distr-columns
Prepare metrics-statistics dataframe for reactable table

metrics-statistics dataframe with *_distr-columns
}
\description{
Prepare metrics-statistics dataframe for reactable table
}
