% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_analysis_tables.R
\name{prepare_selact_stats_for_table}
\alias{prepare_selact_stats_for_table}
\title{Prepare metrics-statistics dataframe for reactable table}
\usage{
prepare_selact_stats_for_table(
  data,
  basin_id = NULL,
  region_id = NULL,
  axis_data = NULL
)
}
\arguments{
\item{data}{metrics-statistics dataframe}

\item{type}{type of level_type to be included in the dataframe, default is c("Région (total)", "Région")}
}
\description{
Prepare metrics-statistics dataframe for reactable table
}
