% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_table_characteristics.R
\name{reactable_style_function}
\alias{reactable_style_function}
\title{Helper function to style reactable table}
\usage{
reactable_style_function(value, other_value)
}
\arguments{
\item{value}{actual value which should be coloured}

\item{other_value}{value to which the actual value should be compared}
}
\value{
styling for reactable row
}
\description{
Helper function to style reactable table
}
\examples{
# inside reactable function:
colDef(
name = "Tronçon",
style = function(value, index) {
  reactable_style_function(value, merged$region_mean[index])
})
}
