% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_region_hydrographique.R
\docType{data}
\name{region_hydrographique}
\alias{region_hydrographique}
\title{region_hydrographique}
\format{
A data frame with 1 rows and 7 variables:
\describe{
  \item{ gid }{  numeric }
  \item{ cdregionhy }{  character }
  \item{ lbregionhy }{  character }
  \item{ cdbh }{  character }
  \item{ click }{  logical }
  \item{ opacity }{  numeric }
  \item{ geom }{  sfc_MULTIPOLYGON,sfc }
}
}
\source{
Source
}
\usage{
region_hydrographique
}
\description{
Description.
}
\keyword{datasets}
