% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sld_style_fct.R
\name{sld_get_quantile_colors}
\alias{sld_get_quantile_colors}
\title{Get Quantile Colors}
\usage{
sld_get_quantile_colors(
  quantile_breaks = sld_get_quantile_metric(selected_region_id = region_click_id(),
    selected_metric = selected_metric())
)
}
\arguments{
\item{quantile_breaks}{A numeric vector containing quantile breaks.}
}
\value{
A character vector of colors generated based on quantile breaks.
}
\description{
This function generates a color palette based on quantile breaks for data mapping.
}
\examples{
\dontrun{
  # Example usage:
  quantile_colors <- sld_get_quantile_colors(quantile_breaks = c(0.1, 0.3, 0.5, 0.7, 0.9))
}

}
