% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sld_style_fct.R
\name{sld_get_quantile_metric}
\alias{sld_get_quantile_metric}
\title{Get Quantile Metrics for a Selected Region}
\usage{
sld_get_quantile_metric(selected_region_id, selected_metric)
}
\arguments{
\item{selected_region_id}{The ID of the selected region.}

\item{selected_metric}{The name of the selected metric.}
}
\value{
A numeric vector containing quantile metrics (Q1, Q2, Q3, Q4, Q5) for the selected metric within the specified region.
}
\description{
This function calculates quantile metrics (Q1, Q2, Q3, Q4, Q5) for a selected metric within a specified region.
}
\examples{
# get quantiles from active_channel_width metric
quantile_metrics <- sld_get_quantile_metric(selected_region_id = 11,
                                            selected_metric = "active_channel_width")
quantile_metrics

}
