% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sld_style_fct.R
\name{sld_get_style}
\alias{sld_get_style}
\title{Get Styled Layer Descriptor (SLD) for network metric layer}
\usage{
sld_get_style(
  breaks = sld_get_quantile_metric(selected_region_id = region_click_id(),
    selected_metric = selected_metric()),
  colors = sld_get_quantile_colors(quantile_breaks =
    sld_get_quantile_metric(selected_region_id = region_click_id(), selected_metric =
    selected_metric())),
  metric = selected_metrics()
)
}
\arguments{
\item{breaks}{A numeric vector containing quantile breaks.}

\item{colors}{A character vector of colors generated based on quantile breaks.}

\item{metric}{The name of the selected metric.}
}
\value{
A character string containing the SLD XML for styling data visualization.
}
\description{
This function generates a Styled Layer Descriptor (SLD) XML for styling network metric layer based on quantile breaks colors.
}
\examples{
\dontrun{
  # Example usage:
  sld_style <- sld_get_style(breaks = c(0.1, 0.3, 0.5, 0.7, 0.9),
                             colors = c("green", "yellow", "orange", "red", "blue"),
                             metric = "some_metric")
}

}
