% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_classes_manual.R
\name{sld_get_style_legend}
\alias{sld_get_style_legend}
\title{Get Styled Layer Descriptor (SLD) for legend of manually classified network_metrics layer}
\usage{
sld_get_style_legend(breaks, colors, metric)
}
\arguments{
\item{breaks}{A numeric vector containing quantile breaks.}

\item{colors}{A character vector of colors generated based on quantile breaks.}

\item{metric}{The name of the selected metric.}
}
\value{
A character string containing the SLD XML for styling data visualization.
}
\description{
This function generates a Styled Layer Descriptor (SLD) XML for the legend of the network metric layer which is manually classified based on quantile breaks colors.
}
