% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot_long.R
\name{smoothen_classes}
\alias{smoothen_classes}
\title{Smoothing of classified df}
\usage{
smoothen_classes(df, n_smooth)
}
\arguments{
\item{df}{classified axis data}

\item{n_smooth}{integer number of values to smooth out}
}
\value{
df_smooth smoothed classified axis data
}
\description{
Smoothing of classified df
}
