% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_track.R
\name{track_inputs}
\alias{track_inputs}
\title{Track input user inputs.}
\usage{
track_inputs(input = input)
}
\arguments{
\item{input}{reactivesValues Shiny input.}
}
\value{
message with the inputs name and value as json format.
}
\description{
Track input user inputs.
}
