% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_track.R
\name{track_session}
\alias{track_session}
\title{Track input user session id and datetime.}
\usage{
track_session(session = session)
}
\arguments{
\item{session}{reactivesValues Shiny input.}
}
\value{
message with the session info name and value as json format.
}
\description{
Track input user session id and datetime.
}
