% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_tracks.R
\name{tracks}
\alias{tracks}
\title{Track input user session.}
\usage{
tracks(input = input)
}
\arguments{
\item{input}{reactivesValues Shiny input.}
}
\value{
message with the list avec the input values tracked.
}
\description{
Track input user session.
}
