% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utile_fct.R
\name{utile_get_category_name}
\alias{utile_get_category_name}
\title{Get the category name of a selected metric}
\usage{
utile_get_category_name(selected_metric)
}
\arguments{
\item{selected_metric}{The name of the metric for which you want to find the
category name.}
}
\value{
The category name to which the selected metric belongs.
}
\description{
This function takes a selected metric and returns the category name to which
the metric belongs. It does this by comparing the selected metric to the
metric names in the categories provided by the `params_metrics_choice` function.
}
\examples{
selected_metric <- "some_metric"
category_name <- utile_get_category_name(selected_metric)

}
