% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utile_fct.R
\name{utile_get_metric_name}
\alias{utile_get_metric_name}
\title{Get the name of a selected metric}
\usage{
utile_get_metric_name(selected_metric)
}
\arguments{
\item{selected_metric}{The name of the metric for which you want to retrieve
the metric name.}
}
\value{
The name of the selected metric.
}
\description{
This function takes a selected metric and returns its corresponding name.
It first determines the category name of the metric using the
`utile_get_category_name` function and then retrieves the metric name
from the category's metrics list provided by the `params_metrics_choice` function.
}
