% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utile_fct.R
\name{utile_get_metric_name_value}
\alias{utile_get_metric_name_value}
\title{Get a named vector of all the metric from the metric type from the params metric list.}
\usage{
utile_get_metric_name_value(metric_type)
}
\arguments{
\item{metric_type}{A character with metric type.}
}
\value{
A named character vector with all the metric names and values.
}
\description{
This function extracts metric names and value with metric type stored in params metric list.
}
\examples{
metrics <- utile_get_metric_name_value("largeur")

}
