% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_utils.R
\name{utils_button_label_with_popover}
\alias{utils_button_label_with_popover}
\title{Create buttons with popover for metric labels.}
\usage{
utils_button_label_with_popover(metric_type)
}
\arguments{
\item{metric_type}{A character string specifying the metric type.}
}
\value{
A list of buttons with popovers.
}
\description{
This function generates a list of buttons with popovers that display additional
information for each metric label.
}
\details{
This function uses the `params_metrics_choice` data structure to
extract metric labels and corresponding popover information for the specified
metric type.
}
\examples{
metric_buttons <- utils_button_label_with_popover("largeur")

}
