% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_utils.R
\name{utils_is_identical_or_null}
\alias{utils_is_identical_or_null}
\title{Check if two objects are identical or if one is NULL}
\usage{
utils_is_identical_or_null(x, y)
}
\arguments{
\item{x}{First object to compare. Can be any R object.}

\item{y}{Second object to compare. Can be any R object.}
}
\value{
A logical value (`TRUE` or `FALSE`). Returns `TRUE` if either `x` or `y` is `NULL` or if they are identical, otherwise `FALSE`.
}
\description{
This function compares two objects and returns `TRUE` if either is `NULL` or if both are identical.
}
\examples{
is_identical_or_null(NULL, 1)       # TRUE
is_identical_or_null(5, 5)          # TRUE
is_identical_or_null(5, 6)          # FALSE
is_identical_or_null(NULL, NULL)    # TRUE
}
