% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_utils.R
\name{utils_is_vector_identical}
\alias{utils_is_vector_identical}
\title{Check if two vectors are identical in length and content}
\usage{
utils_is_vector_identical(x, y)
}
\arguments{
\item{x}{First vector to compare.}

\item{y}{Second vector to compare.}
}
\value{
A logical value (`TRUE` or `FALSE`). Returns `TRUE` if both vectors have the same length and the same elements, otherwise `FALSE`.
}
\description{
This function compares two vectors and returns `TRUE` if they have the same length and the same elements (irrespective of order).
}
\examples{
is_vector_identical(c(1, 2, 3), c(3, 2, 1))         # TRUE
is_vector_identical(c(1, 2), c(1, 2, 3))            # FALSE
is_vector_identical(c("a", "b"), c("a", "b"))       # TRUE
is_vector_identical(c("a", "b"), c("a", "b", "c"))  # FALSE
}
