% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data.R
\name{assign_classes_proposed}
\alias{assign_classes_proposed}
\title{Assign classes to network dgos}
\usage{
assign_classes_proposed(data, proposed_class)
}
\arguments{
\item{data}{dataframe or sf object containing all dgos of an axis or region}

\item{proposed_class}{string indicating the type of classification which should be applied to the data}
}
\value{
classified dataframe/sf object with additional variables: class_name and colors
}
\description{
Assign classes to network dgos
}
\examples{
classified_network <- network_dgo \%>\%
    assign_classes(proposed_class = "class_strahler")

}
