% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot_crosssection.R
\name{cr_profile_main}
\alias{cr_profile_main}
\title{Create a cross section plot for selected DGO data.}
\usage{
cr_profile_main(data, axis_toponyme)
}
\arguments{
\item{data}{data.frame elevation profiles from selected dgo.}

\item{axis_toponyme}{text axis toponyme name.}
}
\value{
plotly cross section plot.
}
\description{
Create a cross section plot for selected DGO data.
}
