% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_analysis_plots.R
\name{create_analysis_biplot}
\alias{create_analysis_biplot}
\title{Create a Biplot with Linear Regression and Statistical Annotations}
\usage{
create_analysis_biplot(
  df,
  metric_x,
  metric_y,
  classes = FALSE,
  lm = FALSE,
  axis_name
)
}
\arguments{
\item{df}{A data frame containing the metrics to be plotted.}

\item{metric_x}{A string specifying the name of the x-axis metric (column in the dataframe).}

\item{metric_y}{A string specifying the name of the y-axis metric (column in the dataframe).}

\item{classes}{A boolean indicating whether to group the data by a categorical variable.}

\item{lm}{A boolean indicating whether to add a linear regression line.}
}
\value{
A `plotly` plot object showing the biplot, regression line, and statistical annotations.
}
\description{
This function creates a biplot using the `plotly` library that displays a scatterplot
of two metrics from a given dataframe, fits a linear regression line, and shows key
statistical measures (correlation, R², p-value) in a floating text box.
}
\examples{
create_analysis_biplot(df = mtcars, metric_x = "mpg", metric_y = "wt", classes = FALSE, lm = TRUE)
}
