% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data.R
\name{create_df_input}
\alias{create_df_input}
\title{Create initial dataframe for 1-variable classification, to be displayed on the UI table}
\usage{
create_df_input(axis_data, variable_name, no_classes = 4, quantile = 95)
}
\arguments{
\item{axis_data}{sf-object of an axis, containing all dgos inside the axis}

\item{variable_name}{name of variable for which the classification should be undertaken}

\item{no_classes}{number of classes to be generated}

\item{quantile}{size of quantile which provides value-range of classification}
}
\value{
dataframe with 4 columns: class (name of each class, here automatically set from A-Z),
        variable (variable chosen for classification), greaterthan (values defining the threshold of each class),
        and color (defining the coloring for the map)
}
\description{
Create initial dataframe for 1-variable classification, to be displayed on the UI table
}
\examples{
df <- create_df_input(
      axis_data = network_dgo,
      variable_name = input$variable,
      no_classes = input$no_classes,
      quantile = input$quantile
      )

}
