% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data.R
\name{data_get_axis_dgos_from_region}
\alias{data_get_axis_dgos_from_region}
\title{Get Network Metrics Data for a Specific Region}
\usage{
data_get_axis_dgos_from_region(selected_region_id, con)
}
\arguments{
\item{con}{Connection to Postgresql database.}

\item{selected_axis_id}{The ID of the selected region.}
}
\value{
A sf data frame containing information about network metrics for the specified region.
}
\description{
This function retrieves data about network metrics for a specific region based on its ID.
}
\examples{
con <- db_con()
network_metrics_data <- data_get_axis_dgos(selected_region_id = 33, con = con)
DBI::dbDisconnect(con)

}
