% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_fct.R
\name{data_get_axis_start_end}
\alias{data_get_axis_start_end}
\title{Get the start and end coordinates of a spatial object's axis}
\usage{
data_get_axis_start_end(dgo_axis = dgo_axis())
}
\arguments{
\item{dgo_axis}{A spatial object with a LINESTRING geometry representing an axis.}
}
\value{
A data frame with two rows, where the first row contains the start
        coordinates (x and y) and the second row contains the end coordinates (x and y).
}
\description{
This function takes a spatial object with a LINESTRING geometry and returns
a data frame containing the start and end coordinates of the axis.
}
\examples{
# Create a simple LINESTRING object
line <- st_linestring(matrix(c(0, 0, 1, 1, 2, 2), ncol = 2))
df <- data_get_axis_start_end(line)
print(df)

}
