% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data.R
\name{data_get_bassins}
\alias{data_get_bassins}
\title{Get Hydrographic Basins}
\usage{
data_get_bassins()
}
\value{
A sf data frame containing information about hydrographic basins.
}
\description{
This function retrieves hydrographic basins.
}
\examples{
data <- data_get_bassins()

}
