% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data.R
\name{data_get_elevation_profiles}
\alias{data_get_elevation_profiles}
\title{Get elevation profiles data from selected dgo gid.}
\usage{
data_get_elevation_profiles(selected_dgo_gid)
}
\arguments{
\item{selected_dgo_gid}{integer selected dgo gid}
}
\value{
data.frame
}
\description{
Get elevation profiles data from selected dgo gid.
}
\examples{
data_get_elevation_profiles(selected_dgo_gid = 29567)
}
