% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data.R
\name{data_get_hydro_sites}
\alias{data_get_hydro_sites}
\title{Get hydrometric sites.}
\usage{
data_get_hydro_sites(selected_region_id, con)
}
\arguments{
\item{selected_region_id}{The ID of the selected region.}

\item{con}{PqConnection to Postgresql database.}
}
\value{
sf data frame containing information about the hydrometric sites within the specified region.
}
\description{
This function retrieves data about the hydrometric sites from Hubeau.
}
\examples{
con <- db_con()
hydro_sites <- data_get_hydro_sites(selected_region_id = 11, con = con)
DBI::dbDisconnect(con)

}
