% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_fct.R
\name{data_get_regions_in_bassin}
\alias{data_get_regions_in_bassin}
\title{Get all the hydrological regions in a Hydrographic Basin}
\usage{
data_get_regions_in_bassin(selected_bassin_id = bassin_click$id)
}
\arguments{
\item{selected_bassin_id}{The ID of the selected hydrographic basin.}
}
\value{
A df data frame containing regions within the specified hydrographic basin.
}
\description{
This function retrieves regions within a specified hydrographic basin based on its ID.
}
\examples{
\dontrun{
  # Example usage:
  data <- data_get_regions_in_bassin(selected_bassin_id = 'cdbh_id')
}

}
