% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data.R
\name{data_get_regions_in_bassin}
\alias{data_get_regions_in_bassin}
\title{Get all the hydrological regions in a Hydrographic Basin}
\usage{
data_get_regions_in_bassin(selected_bassin_id, opacity, con)
}
\arguments{
\item{selected_bassin_id}{text ID of the selected hydrographic basin.}

\item{opacity}{list that contain numeric values clickable and not_clickable to inform the user the non available features.}

\item{con}{PqConnection to Postgresql database.}
}
\value{
A df data frame containing regions within the specified hydrographic basin.
}
\description{
This function retrieves regions within a specified hydrographic basin based on its ID.
}
\examples{
con <- db_con()
opacity = list(clickable = 0.01,
               not_clickable = 0.10)
data <- data_get_regions_in_bassin(selected_bassin_id = "06",
                                   opacity = opacity,
                                   con = con)
DBI::dbDisconnect(con)

}
