% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data.R
\name{data_get_station_hubeau}
\alias{data_get_station_hubeau}
\title{Get hydrometric stations}
\usage{
data_get_station_hubeau(selected_region_id)
}
\arguments{
\item{selected_region_id}{The ID of the selected region.}
}
\value{
A sf data frame containing information about the hydrometric stations within the specified region.
}
\description{
This function retrieves data about the hydrometric stations from Hubeau within a specified region based on its ID.
}
\examples{
hydro_stations <- data_get_station_hubeau(selected_region_id = 11)

}
