% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data.R
\name{data_get_station_hubeau}
\alias{data_get_station_hubeau}
\title{get hydrometric station from Hubeau API}
\usage{
data_get_station_hubeau(region_hydro)
}
\arguments{
\item{region_hydro}{sf polygon object.}
}
\value{
sf object
}
\description{
Extract hydrometric station from Hubeau API https://hubeau.eaufrance.fr/page/api-ecoulement#/ecoulement
}
