% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data.R
\name{data_get_station_hubeau}
\alias{data_get_station_hubeau}
\title{get hydrometric station from Hubeau API}
\usage{
data_get_station_hubeau(hubeau_url)
}
\value{
data.frame
}
\description{
Extract hydrometric station from Hubeau API https://hubeau.eaufrance.fr/page/api-ecoulement#/ecoulement
}
\examples{
data <- data_get_station_hubeau("https://hubeau.eaufrance.fr/api/v1/ecoulement/stations?format=json&bbox=1.6194&bbox=47.7965&bbox=2.1910&bbox=47.9988&size=2")

}
