% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_table_characteristics.R
\name{fct_table_create_reactable}
\alias{fct_table_create_reactable}
\title{Create reactable table}
\usage{
fct_table_create_reactable(df, unit)
}
\arguments{
\item{df}{dataframe of merged regional, axis and dgo data}

\item{unit}{}
}
\description{
Create reactable table
}
