% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_table_characteristics.R
\name{fct_table_create_table_df}
\alias{fct_table_create_table_df}
\title{Merge regional, axis, and dgo dataframes together for reactable table}
\usage{
fct_table_create_table_df(region_sf, axis_sf, dgo_sf)
}
\arguments{
\item{region_sf}{sf object of dgos from selected regional network}

\item{axis_sf}{sf object of dgos from selected axis network}

\item{dgo_sf}{sf object of selected dgo}
}
\description{
Merge regional, axis, and dgo dataframes together for reactable table
}
\examples{
fct_table_create_table_df(network_dgo, network_dgo, network_dgo[1,])
}
