% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lg_profile_fct.R
\name{lg_profile_main}
\alias{lg_profile_main}
\title{Create a longitudinal profile plot for selected axis data}
\usage{
lg_profile_main(data, y, y_label, y_label_category)
}
\arguments{
\item{data}{A data frame containing the selected axis data.}

\item{y}{The metric to be plotted on the y-axis.}

\item{y_label}{The name of the metric plotted.}

\item{y_label_category}{The metric category name.}
}
\value{
A longitudinal profile plot with the specified metric.
}
\description{
This function generates a longitudinal profile plot using the 'plot_ly'
function from the 'plotly' package. It allows you to visualize a specific
metric along the selected axis.
}
\examples{
# Create a longitudinal profile plot for active channel width
selected_axis_df <- as.data.frame(network_dgo)

}
