% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lg_profile_fct.R
\name{lg_profile_second}
\alias{lg_profile_second}
\title{Create a dual-axis longitudinal profile plot for selected axis data}
\usage{
lg_profile_second(data, y, y_label, y_label_category)
}
\arguments{
\item{data}{A data frame containing the selected axis data.}

\item{y}{The primary metric to be plotted on the left y-axis.}

\item{y_label}{The name of the metric plotted.}

\item{y_label_category}{The metric category name.}
}
\value{
A dual-axis longitudinal profile plot with the specified metrics.
}
\description{
This function generates a dual-axis longitudinal profile plot using the 'plot_ly'
function from the 'plotly' package. It allows you to visualize two different
metrics along the selected axis.
}
\examples{
\dontrun{
# like lg_profile_update_main function, see example in documentation
}

}
