% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_fct.R
\name{map_add_basemaps}
\alias{map_add_basemaps}
\title{Add Basemap Layers to an Existing Leaflet Map}
\usage{
map_add_basemaps(map, basemaps)
}
\arguments{
\item{map}{An existing Leaflet map to which basemap layers will be added.}

\item{basemaps}{A data frame containing information about the basemap layers to be added.}
}
\value{
An updated Leaflet map with basemap layers added.
}
\description{
This function adds basemap layers to an existing Leaflet map.
}
\examples{
\dontrun{
  # Example usage:
  updated_map <- map_add_basemaps(map = existing_map, basemaps = basemaps_data)
}

}
