% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_map.R
\name{map_add_basemaps}
\alias{map_add_basemaps}
\title{Add Basemap Layers to an Existing Leaflet Map}
\usage{
map_add_basemaps(map)
}
\arguments{
\item{map}{An existing Leaflet map to which basemap layers will be added.}
}
\value{
An updated Leaflet map with basemap layers added.
}
\description{
This function adds basemap layers to an existing Leaflet map.
}
\examples{
\dontrun{
  # Used in map_init_bassins() function, its use is in the function
}

}
