% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_map.R
\name{map_add_network_metric}
\alias{map_add_network_metric}
\title{Add a metric layer with custom symbology to a map.}
\usage{
map_add_network_metric(
  map,
  wms_params = globals$wms_params$metric,
  sld_body = "",
  group = "Réseau hydrographique"
)
}
\arguments{
\item{map}{A leaflet map object to which the metric layer will be added.}

\item{wms_params}{A list containing WMS parameters for the metric layer. If not provided, default parameters are retrieved using the \code{\link{params_wms}} function.}

\item{sld_body}{A character string representing the SLD (Styled Layer Descriptor) body for custom styling of the metric layer.}

\item{cql_filter}{A character string representing a CQL filter to apply to the metric layer.}

\item{data_axis}{A data axis to display on the map.}
}
\value{
A leaflet map object with the metric layer added.
}
\description{
This function adds a metric layer with custom symbology to a leaflet map. It allows you to specify custom parameters for the Web Map Service (WMS) request, apply a CQL (Common Query Language) filter, and provide a custom SLD (Styled Layer Descriptor) body for styling the layer. Additionally, you can specify the data axis to display on the map.
}
