% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_fct.R
\name{map_add_overlayers}
\alias{map_add_overlayers}
\title{Add Overlayer Layers to an Existing Leaflet Map}
\usage{
map_add_overlayers(map, overlayers)
}
\arguments{
\item{map}{An existing Leaflet map to which overlayer layers will be added.}

\item{overlayers}{A data frame containing information about the overlayer layers to be added.}
}
\value{
An updated Leaflet map with overlayer layers added.
}
\description{
This function adds overlayer layers to an existing Leaflet map.
}
\examples{
\dontrun{
  # Example usage:
  updated_map <- map_add_overlayers(map = existing_map, overlayers = overlayers_data)
}

}
