% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_map.R
\name{map_add_regions_in_bassin}
\alias{map_add_regions_in_bassin}
\title{Add hydrological Regions in a Bassin to an existing Leaflet Map}
\usage{
map_add_regions_in_bassin(
  map,
  bassins_data,
  bassin_click = bassin_click,
  regions_data = region_hydro
)
}
\arguments{
\item{map}{An existing Leaflet map to which regions will be added.}

\item{bassins_data}{A sf data.frame with bassins data.}

\item{bassin_click}{A vector containing information about the clicked bassin.}

\item{regions_data}{A sf data.frame containing information about regions within the bassin.}
}
\value{
An updated Leaflet map with regions added.
}
\description{
This function adds regions within a bassin to an existing Leaflet map.
}
\examples{
library(leaflet)
library(dplyr)
library(sf)
# Create init bassin map
my_map <- map_init_bassins(bassins_data = bassin_hydrographique,
                           id_logo_ign_remonterletemps = "logo_ign_remonterletemps")

# simulate bassin selected
selected_bassin <- bassin_hydrographique

# get centroid coordinate (in shiny see leaflet mapid_shape_click)
centre <- sf::st_centroid(selected_bassin)
centre_coord <- as.data.frame(st_coordinates(centre)) \%>\%
  rename("lng" = X,
         "lat" = Y)
# map region
map <- map_add_regions_in_bassin(map = my_map,
                                 bassins_data = bassin_hydrographique,
                                 bassin_click = centre_coord,
                                 regions_data = region_hydrographique)
map

}
