% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_fct.R
\name{map_init_bassins}
\alias{map_init_bassins}
\title{Initialize a Leaflet Map with hydrological Bassins}
\usage{
map_init_bassins(bassins_data, id_logo_ign_remonterletemps)
}
\arguments{
\item{bassins_data}{A hydrological bassins sf data frame.}

\item{id_logo_ign_remonterletemps}{id of the IGN remonter le temps image.}
}
\value{
A Leaflet map object with basemaps, scale and layer control.
}
\description{
This function initializes a Leaflet map with bassins data and various map layers.
}
\examples{
map <- map_init_bassins(bassins_data = bassin_hydrographique,
                        id_logo_ign_remonterletemps = "logo_ign_remonterletemps")
map

}
