% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_fct.R
\name{map_init_bassins}
\alias{map_init_bassins}
\title{Initialize a Leaflet Map with hydrological Bassins}
\usage{
map_init_bassins(bassins_data = get_bassins())
}
\arguments{
\item{bassins_data}{A hydrological bassins sf data frame.}
}
\value{
A Leaflet map object with basemaps, scale and layer control.
}
\description{
This function initializes a Leaflet map with bassins data and various map layers.
}
\examples{
\dontrun{
  # Example usage:
  map <- map_init_bassins(bassins_data = some_bassins_data)
}

}
