% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_fct.R
\name{map_legend_metric}
\alias{map_legend_metric}
\title{Generate and display a legend for a map layer using a provided SLD (Styled Layer Descriptor) body.}
\usage{
map_legend_metric(sld_body)
}
\arguments{
\item{sld_body}{A character string containing the SLD (Styled Layer Descriptor) body that defines the map layer's styling.}
}
\value{
An HTML img tag representing the legend for the specified map layer.
}
\description{
This function constructs a legend for a map layer by sending a GetLegendGraphic request to a WMS (Web Map Service) server.
}
\examples{
# Define an SLD body for a map layer
sld_body <- '<StyledLayerDescriptor>...</StyledLayerDescriptor>'

# Generate and display the legend for the map layer
legend <- map_legend_metric(sld_body)
print(legend)

}
