% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_map.R
\name{map_legend_metric}
\alias{map_legend_metric}
\title{Generate and display a legend for a map layer using a provided SLD (Styled Layer Descriptor) body.}
\usage{
map_legend_metric(sld_body)
}
\arguments{
\item{sld_body}{A character string containing the SLD (Styled Layer Descriptor) body that defines the map layer's styling.}
}
\value{
An HTML img tag representing the legend for the specified map layer.
}
\description{
This function constructs a legend for a map layer by sending a GetLegendGraphic request to a WMS (Web Map Service) server.
}
\examples{
con <- db_con()
# Define an SLD body for a map layer
sld_body <- sld_get_style(breaks = sld_get_quantile_metric(
                                     selected_region_id = 11,
                                     selected_metric = "active_channel_width",
                                     con = con),
                          colors = sld_get_quantile_colors(
                                     quantile_breaks = sld_get_quantile_metric(
                                                         selected_region_id = 11,
                                                         selected_metric = "active_channel_width",
                                                         con = con)),
                          metric = "active_channel_width")
DBI::dbDisconnect(con)

# Generate and display the legend for the map layer
legend <- map_legend_metric(sld_body)

}
