% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_fct.R
\name{map_metric}
\alias{map_metric}
\title{Update metric mapping}
\usage{
map_metric(
  map_id = "exploremap",
  data_map = network_filter(),
  varsel = varsel(),
  network_group = "D",
  data_axis = network_region_axis(),
  axis_group = "AXIS"
)
}
\arguments{
\item{map_id}{map shiny id}

\item{data_map}{network data}

\item{varsel}{metric selected data}

\item{network_group}{network group}

\item{data_axis}{axis data}

\item{axis_group}{axis group}
}
\value{
leaflet update
}
\description{
Update metric mapping
}
\examples{
map_metric(map_id = "exploremap", data_map = network_filter(), varsel = varsel(),
network_group = "D", data_axis = network_region_axis(), axis_group = "AXIS")
}
