% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_fct.R
\name{map_no_metric}
\alias{map_no_metric}
\title{Update map without metric selected}
\usage{
map_no_metric(
  map,
  geoserver_url,
  network_metrics_wms,
  wms_format,
  metric_group = "METRIC",
  selected_region_id = selected_region_feature()[["gid"]],
  strahler_filter_min = input$strahler[1],
  strahler_filter_max = input$strahler[2],
  data_axis,
  axis_group = "AXIS"
)
}
\arguments{
\item{map}{map to update}

\item{data_axis}{axis data}

\item{axis_group}{axis layer group}
}
\value{
update leaflet map
}
\description{
Update map without metric selected
}
\examples{
leafletProxy("exploremap") \%>\%
  map_no_metric(data_network = network_filter(),  network_group = "D", data_axis = network_region_axis(), axis_group = "AXIS")
}
