% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_fct.R
\name{map_region_clicked}
\alias{map_region_clicked}
\title{Map region click}
\usage{
map_region_clicked(
  map,
  region_click = region_click,
  selected_region_feature = selected_region_feature,
  regions_group = "REGION",
  selected_region_group = "SELECT_REGION"
)
}
\arguments{
\item{map}{map object for pipe}

\item{region_click}{region clicked parameter app}

\item{selected_region_feature}{region clicked feature}

\item{regions_group}{regions group to clear}

\item{selected_region_group}{selected regions group}
}
\value{
leaflet update
}
\description{
Map region click
}
\examples{
map_region_clicked(region_click = region_click,
selected_region_feature = selected_region_feature,
regions_group = "REGION",
selected_region_group = "SELECT_REGION")
}
