% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_params.R
\name{params_map_group}
\alias{params_map_group}
\title{Get Parameters for Map Layer Groups}
\usage{
params_map_group()
}
\value{
A list of parameters including names for groups such as "BASSIN," "REGION," "SELECT_REGION," "METRIC," "AXIS," "LEGEND," and "ROE".
}
\description{
This function returns a list of parameters representing different map layer groups.
}
\examples{
# all group available
map_group_params <- params_map_group()
# get specific group name
map_metric_group <- params_map_group()$metric
map_selected_region_group <- params_map_group()$select_region

}
