% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params_fct.R
\name{params_metrics_choice}
\alias{params_metrics_choice}
\title{Get Choices for Metric Selection}
\usage{
params_metrics_choice()
}
\value{
A list of choices for selecting metrics, categorized into "Largeurs (m)," "Pentes," "Occupation du sol," "Continuité latérale," and "Indices."
}
\description{
This function returns a list of choices for selecting metrics organized into categories.
}
\examples{
\dontrun{
  # Example usage:
  metric_choices <- params_metrics_choice()
}

}
