% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_analysis_tables.R
\name{prepare_regions_stats_for_table}
\alias{prepare_regions_stats_for_table}
\title{Prepare metrics-statistics dataframe for reactable table for regions}
\usage{
prepare_regions_stats_for_table(data, region_names = NULL)
}
\arguments{
\item{data}{metrics-statistics dataframe}

\item{region_names}{df with ids and names of regions}
}
\description{
Prepare metrics-statistics dataframe for reactable table for regions
}
