% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_analysis_tables.R
\name{prepare_stats_df}
\alias{prepare_stats_df}
\title{Prepare metrics-statistics dataframe for reactable table}
\usage{
prepare_stats_df(data, type = c("Région (total)", "Région"), region_names)
}
\arguments{
\item{data}{metrics-statistics dataframe}

\item{type}{type of level_type to be included in the dataframe, default is c("Région (total)", "Région")}
}
\value{
metrics-statistics dataframe with *_distr-columns
}
\description{
Prepare metrics-statistics dataframe for reactable table
}
