% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_sld_style.R
\name{sld_get_quantile_colors}
\alias{sld_get_quantile_colors}
\title{Get Quantile Colors}
\usage{
sld_get_quantile_colors(quantile_breaks)
}
\arguments{
\item{quantile_breaks}{A numeric vector containing quantile breaks.}
}
\value{
A character vector of colors generated based on quantile breaks.
}
\description{
This function generates a color palette based on quantile breaks for data mapping.
}
\examples{
# get quantiles from active_channel_width metric
quantile_metrics <- sld_get_quantile_metric(selected_region_id = 11,
                                            selected_metric = "active_channel_width")
# get color from quantile
quantile_colors <- sld_get_quantile_colors(quantile_breaks = quantile_metrics)
quantile_colors

}
