% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sld_style_fct.R
\name{sld_get_style}
\alias{sld_get_style}
\title{Get Styled Layer Descriptor (SLD) for network metric layer}
\usage{
sld_get_style(breaks, colors, metric)
}
\arguments{
\item{breaks}{A numeric vector containing quantile breaks.}

\item{colors}{A character vector of colors generated based on quantile breaks.}

\item{metric}{The name of the selected metric.}
}
\value{
A character string containing the SLD XML for styling data visualization.
}
\description{
This function generates a Styled Layer Descriptor (SLD) XML for styling network metric layer based on quantile breaks colors.
}
\examples{
# get quantiles from active_channel_width metric
quantile_metrics <- sld_get_quantile_metric(selected_region_id = 11,
                                            selected_metric = "active_channel_width")
# get color from quantile
quantile_colors <- sld_get_quantile_colors(quantile_breaks = quantile_metrics)
# create sld style
sld_style <- sld_get_style(breaks = quantile_metrics,
                           colors = quantile_colors,
                           metric = "active_channel_width")
sld_style

}
